Rename dash to underdash
