+++ src/collection.js
@@ -1,5 +1,5 @@
 import Events from './events.js';
-import _ from './utils/dash.js';
+import _ from './utils/underdash.js';
 
 
 /**
