+++ src/effects/image-sequence.js
@@ -2,7 +2,7 @@ import Collection from '../collection.js';
 import Events from '../events.js';
 import defineGetterSetter from '../utils/get-set.js';
 import {lerp} from '../utils/math.js';
-import _ from '../utils/dash.js';
+import _ from '../utils/underdash.js';
 
 import Path from '../path.js';
 import Anchor from '../anchor.js';
