+++ src/effects/sprite.js
@@ -1,6 +1,6 @@
 import defineGetterSetter from '../utils/get-set.js';
 import {lerp} from '../utils/math.js';
-import _ from '../utils/dash.js';
+import _ from '../utils/underdash.js';
 
 import Path from '../path.js';
 import Anchor from '../anchor.js';
