+++ src/shapes/rectangle.js
@@ -1,7 +1,7 @@
 import Commands from '../utils/path-commands.js';
 import Events from '../events.js';
 import defineGetterSetter from '../utils/get-set.js';
-import _ from '../utils/dash.js';
+import _ from '../utils/underdash.js';
 
 import Path from '../path.js';
 import Anchor from '../anchor.js';
