+++ src/text.js
@@ -1,7 +1,7 @@
 import {getComputedMatrix} from './utils/math.js';
 import Events from './events.js';
 import defineGetterSetter from './utils/get-set.js';
-import _ from './utils/dash.js';
+import _ from './utils/underdash.js';
 
 import Shape from './shape.js';
